New System native class
