+++ c/optionals/optionals.h
@@ -3,10 +3,9 @@
 
 #include "math.h"
 #include "env.h"
+#include "system.h"
 #include "../vm.h"
 
 void defineNativeMethod(ObjClassNative *klass, const char *name, NativeFn function);
-
-void defineNativeVoidMethod(ObjClassNative *klass, const char *name, NativeFnVoid function);
 
 #endif //dictu_optionals_h
