+++ docs/docs/env.md
@@ -20,7 +20,7 @@ nav_order: 12
 You can access ENV vars through the use of the Env.get() method. If the ENV var does not exist nil is returned, other wise
 a string value is returned.
 
-```
+```js
 Env.get("bad key!"); // nil
 Env.get("valid key"); // "value"
 ```
@@ -30,7 +30,7 @@ Env.get("valid key"); // "value"
 You can set ENV vars through the use of the Env.set() method. You can also clear an ENV var by passing a `nil` value.
 When setting an ENV var the key must be a string and the value must be either a string or nil;
 
-```
+```js
 Env.set("key", "test");
 Env.set("key", nil); // Remove env var
 Env.set("key", 10); // set() arguments must be a string or nil.
