+++ tests/builtins/import.du
@@ -8,4 +8,3 @@ import "tests/builtins/str.du";
 import "tests/builtins/number.du";
 import "tests/builtins/type.du";
 import "tests/builtins/len.du";
-import "tests/builtins/sleep.du";
