--- tests/builtins/sleep.du
@@ -1,13 +1,0 @@
-/**
-* sleep.du
-*
-* Testing the builtin sleep() function
-*
-* sleep() pauses execution for a given amount of seconds
-*/
-
-var start = time();
-
-sleep(2);
-
-assert(time() - start == 2);
