+++ tests/runTests.du
@@ -10,6 +10,7 @@ import "tests/builtins/import.du";
 import "tests/maths/import.du";
 import "tests/loops/import.du";
 import "tests/env/import.du";
+import "tests/system/import.du";
 
 // If we got here no runtime errors were thrown, therefore all tests passed.
 print("All tests passed successfully!");
