+++ tests/system/import.du
@@ -1,0 +1,7 @@
+/**
+ * import.du
+ *
+ * General import file for all the System methods
+ */
+
+import "tests/system/sleep.du";
